Face Data/requirements.txt
streamlit==1.30.0
torch==2.0.1
torchvision==0.15.2
pillow==10.0.0
numpy==1.24.3
matplotlib==3.7.2
opencv-python==4.8.0.76
scikit-learn>=0.24.0
jupyter>=1.0.0
tqdm>=4.62.0
requests>=2.25.0
google-generativeai>=0.1.0  # For Gemini API
groq>=0.3.0  # For Groq API
ipywidgets>=7.6.0 
scipy==1.10.1 